Data.Type.Combinator.Singletons

Description

There's a lot of identical data types between type-combinators and singetons, as well as similar typeclasses. This module provides conversion functions between the two (through the TC typeclass), and also many appropriate orphan instances.

Conversion functions

class TC f where

Typeclass for type-combinator types that can be converted to and from singletons.

Minimal complete definition

fromTC, toTC

Methods

fromTC :: f a -> Sing a

Convert a type-combinator type that is equivalent to a singleton into its equivalent Sing.

toTC :: Sing a -> f a

Convert a Sing into its equivalent type-combinator type.

singLength :: Sing as -> Length as

Convert a Sing for as into a Length representing the length of as.

Length as is equivalent to Prod Proxy as, so this is basically

singLength :: Sing as -> Prod Proxy as
singLength = map1 (const Proxy) . toTC

This function is one-way, since the actual run-time information on the types in as is lost.

singSome :: SomeSing k -> Some (Sing :: k -> Type)

Convert SomeSing k from singletons into the more generic representation Some Sing, from type-combinators.

someSing :: Some (Sing :: k -> Type) -> SomeSing k

Convert the generic representation Some Sing from type-combinatprs to SomeSing k, from singletons.

singWit1 :: SingInstance a -> Wit1 SingI a

Convert a SingInstance a from singletons into the more generic representation Wit1 SingI a.

wit1Sing :: Wit1 SingI a -> SingInstance a

Convert the generic representation Wit1 SIngI a from type-combinators to SingInstance a from singletons.

Orphan singleton instance for N

data family Sing (a :: k) :: *

The singleton kind-indexed data family.

type SN = (Sing :: N -> Type)

type ZSym0 = Z

data SSym0 (l :: TyFun N N)

type SSym1 (t :: N) = S t

Orphan instances

SingKind N

Instance details

Associated Types

type Demote N = (r :: *)

Methods

fromSing :: Sing a -> Demote N

toSing :: Demote N -> SomeSing N

SingI Z

Instance details

Methods

sing :: Sing Z

SingI n => SingI (S n :: N)

Instance details

Methods

sing :: Sing (S n)

SEq k => Eq1 (Sing :: k -> *)

Instance details

Methods

eq1 :: Sing a -> Sing a -> Bool

neq1 :: Sing a -> Sing a -> Bool

SOrd k => Ord1 (Sing :: k -> *)

Instance details

Methods

compare1 :: Sing a -> Sing a -> Ordering

(<#) :: Sing a -> Sing a -> Bool

(>#) :: Sing a -> Sing a -> Bool

(<=#) :: Sing a -> Sing a -> Bool

(>=#) :: Sing a -> Sing a -> Bool

SingI a => Known (Sing :: k -> *) (a :: k)

Instance details

Associated Types

type KnownC Sing a :: Constraint

Methods

known :: Sing a

Witness ØC (SingI a) (Sing a)

Instance details

Associated Types

type WitnessC ØC (SingI a) (Sing a) :: Constraint

Methods

(\\) :: ØC => (SingI a -> r) -> Sing a -> r